refactor: concurrency improvements (#68)
* refactor: volume.Delete() do not remove seed snapshot
* refactor: do not repeatedly delete FS
* fix: volume.getUnderlyingSnapshots() failed if volume was already gone
* fix: more precise logging timestamp RFC3339Nano
* feat: configurable concurrenty limitation per operation
* docs: update values.schema.json in CI flow

+++ pkg/wekafs/volume.go
@@ -162,9 +162,10 @@ func (v *Volume) hasUnderlyingSnapshots(ctx context.Context) (bool, error) {
 	logger := log.Ctx(ctx)
 	has := false
 
-	if v.isOnSnapshot() {
+	if !v.isFilesystem() {
 		return false, nil
 	}
+
 	snapshots, err := v.getUnderlyingSnapshots(ctx)
 	if err != nil {
 		return true, err
@@ -195,30 +196,47 @@ func (v *Volume) hasUnderlyingSnapshots(ctx context.Context) (bool, error) {
 func (v *Volume) isAllowedForDeletion(ctx context.Context) bool {
 	if !v.isFilesystem() {
 		return true
-	} else {
-		hasSnaps, err := v.hasUnderlyingSnapshots(ctx)
-		if err != nil {
-			return true // we want to be on the safe side. If we fail to get snapshots, just block the FS from deletion
-		}
-		if hasSnaps {
-			return false
-		}
 	}
-	return true
+	op := "isAllowedForDeletion"
+	ctx, span := otel.Tracer(TracerName).Start(ctx, op)
+	defer span.End()
+	ctx = log.With().Str("trace_id", span.SpanContext().TraceID().String()).Str("span_id", span.SpanContext().SpanID().String()).Str("op", op).Logger().WithContext(ctx)
+	logger := log.Ctx(ctx).With().Str("volume_id", v.GetId()).Str("filesystem", v.FilesystemName).Logger()
+	logger.Trace().Msg("Checking if deletion of volume is allowed")
+	hasSnaps, err := v.hasUnderlyingSnapshots(ctx)
+	if err != nil {
+		logger.Error().Err(err).Msg("Failed to fetch underlying snapshots, filesystem deletion not allowed")
+		return false // we want to be on the safe side. If we fail to get snapshots, just block the FS from deletion
+	}
+	return !hasSnaps
 }
 
 // getUnderlyingSnapshots intended to return list of Weka snapshots that exist for filesystem
 func (v *Volume) getUnderlyingSnapshots(ctx context.Context) (*[]apiclient.Snapshot, error) {
+	op := "getUnderlyingSnapshots"
+	ctx, span := otel.Tracer(TracerName).Start(ctx, op)
+	defer span.End()
+	ctx = log.With().Str("trace_id", span.SpanContext().TraceID().String()).Str("span_id", span.SpanContext().SpanID().String()).Str("op", op).Logger().WithContext(ctx)
+	logger := log.Ctx(ctx).With().Str("volume_id", v.GetId()).Str("filesystem", v.FilesystemName).Logger()
+
 	snapshots := &[]apiclient.Snapshot{}
 	if v.apiClient == nil {
 		return nil, errors.New("cannot check for underlying snaphots as volume is not bound to API")
 	}
+
 	fsObj, err := v.getFilesystemObj(ctx)
 	if err != nil {
 		return nil, err
 	}
 	if fsObj == nil {
-		return nil, errors.New("could not fetch volume snapshots")
+		// filesystem is already deleted or not exists
+		logger.Warn().Msg("Filesystem not exists")
+		return snapshots, nil
+	}
+
+	if fsObj.IsRemoving {
+		// assume snapshots are not relevant in such case
+		return snapshots, nil
 	}
 
 	err = v.apiClient.FindSnapshotsByFilesystem(ctx, fsObj, snapshots)
@@ -1213,13 +1231,6 @@ func (v *Volume) Create(ctx context.Context, capacity int64) error {
 		logger.Debug().Msg("Successully created directory")
 	}
 
-	// create the seed snapshot for the filesystem if needed
-	if v.isFilesystem() && v.server.getConfig().allowAutoSeedSnapshotCreation {
-		if _, err := v.ensureSeedSnapshot(ctx); err != nil {
-			logger.Error().Err(err).Msg("Failed to create seed snapshot, new snapshot volumes cannot be created from this filesystem!")
-		}
-	}
-
 	// Update volume capacity
 	if err := v.UpdateCapacity(ctx, &(v.enforceCapacity), capacity); err != nil {
 		logger.Error().Str("volume_type", string(v.GetType())).
@@ -1314,7 +1325,6 @@ func (v *Volume) Delete(ctx context.Context) error {
 		if !v.isAllowedForDeletion(ctx) {
 			return ErrFilesystemHasUnderlyingSnapshots
 		}
-		v.deleteSeedSnapshot(ctx)
 		err = v.deleteFilesystem(ctx)
 	} else if v.isOnSnapshot() {
 		err = v.deleteSnapshot(ctx)
@@ -1362,26 +1372,29 @@ func (v *Volume) deleteFilesystem(ctx context.Context) error {
 		// FS doesn't exist already, return OK for idempotence
 		return nil
 	}
-	fsUid := fsObj.Uid
-	logger.Trace().Str("filesystem", v.FilesystemName).Msg("Attempting deletion of filesystem")
-	fsd := &apiclient.FileSystemDeleteRequest{Uid: fsObj.Uid}
-	err = v.apiClient.DeleteFileSystem(ctx, fsd)
-	if err != nil {
-		if err == apiclient.ObjectNotFoundError {
-			logger.Debug().Str("filesystem", v.FilesystemName).Msg("Filesystem not found, assuming repeating request")
-			return nil
+	if !fsObj.IsRemoving { // if filesystem is already removing, just wait
+		logger.Trace().Str("filesystem", v.FilesystemName).Msg("Attempting deletion of filesystem")
+		fsd := &apiclient.FileSystemDeleteRequest{Uid: fsObj.Uid}
+		err = v.apiClient.DeleteFileSystem(ctx, fsd)
+		if err != nil {
+			if err == apiclient.ObjectNotFoundError {
+				logger.Debug().Str("filesystem", v.FilesystemName).Msg("Filesystem not found, assuming repeating request")
+				return nil
+			}
+			if _, ok := err.(*apiclient.ApiNotFoundError); ok {
+				logger.Debug().Str("filesystem", v.FilesystemName).Msg("Filesystem not found, assuming repeating request")
+				return nil
+			}
+			if _, ok := err.(*apiclient.ApiBadRequestError); ok {
+				logger.Trace().Err(err).Msg("Bad request during filesystem deletion, probably already removed")
+				return nil
+			}
+			logger.Error().Err(err).Str("filesystem", v.FilesystemName).Msg("Failed to delete filesystem")
+			return status.Errorf(codes.Internal, "Failed to delete filesystem %s: %s", v.FilesystemName, err)
+
 		}
-		if _, ok := err.(*apiclient.ApiNotFoundError); ok {
-			logger.Debug().Str("filesystem", v.FilesystemName).Msg("Filesystem not found, assuming repeating request")
-			return nil
-		}
-		if _, ok := err.(*apiclient.ApiBadRequestError); ok {
-			logger.Trace().Err(err).Msg("Bad request during filesystem deletion, probably already removed")
-			return nil
-		}
-		logger.Error().Err(err).Str("filesystem", v.FilesystemName).Msg("Failed to delete filesystem")
-		return status.Errorf(codes.Internal, "Failed to delete filesystem %s: %s", v.FilesystemName, err)
 	}
+	fsUid := fsObj.Uid
 	err, done := v.waitForFilesystemDeletion(ctx, logger, fsUid)
 	if done {
 		return err
